requirements.txt
streamlit
mistralai
python-dotenv
pathlib